Moved inference tests to test_inference file
